+++ picasso/gausslq.py
@@ -162,7 +162,9 @@ def fit_spots(spots):
 
 
 def fit_spots_parallel(spots, asynch=False):
-    n_workers = max(1, int(0.75 * _multiprocessing.cpu_count()))
+    n_workers = min(
+        60, max(1, int(0.75 * _multiprocessing.cpu_count()))
+    ) # Python crashes when using >64 cores
     n_spots = len(spots)
     n_tasks = 100 * n_workers
     spots_per_task = [